extension: implement "in limbo" reqres processing stage
..., reqres stat, as well as various "limbo mode" toggles and shortcuts

"Limbo" is an optional pre-archival-queue stage for finished reqres that are
ready to be archived but, unlike non-limbo reqres, are not to be archived
automatically.

I.e. enabling "limbo mode" in a tab allows you to manually archive or discard
reqres generated by that tab after the page was already loaded without
re-loading it.

Useful in cases when you need to actually look at a page before deciding if
you want to archive it.

E.g., you enable limbo mode, reload the page, notice there were no updates to
the interesting parts of the page, and so you discard all of the reqres newly
generated by that tab via appropriate button in the add-on popup, or via the
new keyboard shortcut.