extension: implement `problematic` reqres tracking
... as well as relevant UI and documentation.

This flag is needed because, normally, browsers provide no indication when some
parts of the page failed to load properly --- they expect you to actually look
at the page with your eyes to notice something looking broken instead --- which
is not a proper way to do this when you want to be sure that the whole page
with all its resources was archived.

To provide such an indicator, pWebArc will now display the number of
problematic reqres in extension's badge.