extension/lib/lutils: abstract away tab narrowing and start using it everywhere
From now on, when the current tab is a state page narrowed to a tab, the
extension popup will show target tab's settings and state, and shortcuts
will work on its data.

So, e.g., now you can do `Alt+Shift+A` from a `state.html?tab=<someId>` page
and it will work as you'd expect.