extension/background/core: CBOR-dump all reqres fields completely raw
This makes the implementation adhere to the stated technical philosophy
more closely.

The dumps will grow in size a bit, but this is negligible, since they are
compressed by default by all the archival methods now.

`wrrarms` can handle all of this properly for a while now.

The removal of all that `document_url` and `origin_url` logic will also
simplify future things.